+++ example/index.js
@@ -2,13 +2,22 @@ import ExifRotate from '../lib/exif-rotate';
 
 const elem = document.getElementById('file_image');
 elem.onchange = (e) => {
-  ExifRotate.getBase64String(e.target.files, {
-    max_size: 700,
-  }, (base64) => {
-    console.log(base64);
-  });
+  Object.keys(e.target.files).forEach((index) => {
+    const reader = new FileReader();
+    reader.onload = (e) => {
+      const img = new Image();
+      img.onload = () => {
+        const base64 = ExifRotate.getBase64String(img, {
+          max_size: 300,
+        });
+        console.log(base64);
 
-  ExifRotate.showPreviewImage(e.target.files, {
-    max_size: 700,
+        ExifRotate.showPreviewImage(img, {
+          max_size: 700,
+        });
+      };
+      img.src = e.target.result;
+    };
+    reader.readAsDataURL(e.target.files[index]);
   });
 };
