+++ src/exif-rotate.js
@@ -1,26 +1,15 @@
-import { default_container_id } from './configs';
-import { init } from './make-canvas';
+import { setCanvas, appendNewImage } from './helper';
 
 export default class ExifRotate {
   /**
   *  @param img {elem}
   *  @param options {object}
   */
-  static showPreviewImage(files, options = {}) {
-    init(files, options)
-    .then((canvas) => {
-      for (let i = 0; i < canvas.length; i += 1) {
-        const base_64 = canvas[i].toDataURL('image/jpeg');
-        const new_img = new Image();
-        new_img.setAttribute('src', base_64);
-        const id = options.container_id ? options.container_id : default_container_id;
-        const container = document.getElementById(id);
-        container.appendChild(new_img);
-      }
-    })
-    .catch((error) => {
-      console.log(error);
-    });
+  static showPreviewImage(img, options = {}) {
+    const canvas = document.createElement('canvas');
+    const ctx = canvas.getContext('2d');
+    setCanvas(img, canvas, ctx, options);
+    appendNewImage(canvas, options);
   }
 
   /**
@@ -28,15 +17,10 @@ export default class ExifRotate {
   *  @param options {object}
   *  @return base 64 data url {string}
   */
-  static getBase64String(files, options = {}, callback) {
-    init(files, options)
-    .then((canvas) => {
-      for (let i = 0; i < canvas.length; i += 1) {
-        callback(canvas[i].toDataURL('image/jpeg'));
-      }
-    })
-    .catch((error) => {
-      console.log(error);
-    });
+  static getBase64String(img, options = {}) {
+    const canvas = document.createElement('canvas');
+    const ctx = canvas.getContext('2d');
+    setCanvas(img, canvas, ctx, options);
+    return canvas.toDataURL('image/jpeg');
   }
 }
